final completion

+++ assets/js/script1.js
@@ -7,15 +7,16 @@ var opener2 = document.getElementById("opener2");
 var entry = document.getElementById("task");
 var sub = document.getElementById("subbtn");
 var h2heading = document.getElementById("taskslistheading");
-var taskboxs = document.querySelectorAll(".task");
-var editbtn = document.querySelectorAll('.edit');
+
 var body = document.body;
 var bgcolor;
 var colorflag = "white";
-var input_elements = []
-var edit_elements = [];
-var box_elements = [];
 var edit;
+var btnclass;
+var inpclass;
+
+
+
 function theme() {
 
     var bgcolor = window.getComputedStyle(body).getPropertyValue('background-color');
@@ -38,18 +39,7 @@ function theme() {
         h2heading.style.background = "linear-gradient(90deg, #1CB5E0 0%, #000851 100%)";
         h2heading.style.WebkitBackgroundClip = "text";
         h2heading.style.WebkitTextFillColor = "transparent";
-        for (let i = 0; i < edit_elements.length; i++) {
-            var elements = edit_elements[i];
-            elements.classList.remove('pinkedit');
-            elements.classList.add('blackedit');
-            var box = box_elements[i];
-            box.style.backgroundColor = "rgb(37 39 60)";
-            if (elements.innerText.toLowerCase() == "save") {
-                var inp = input_elements[i];
-                inp.style.color = "#00d9ff";
-            }
-
-        }
+        showtasks();
 
     }
     else {
@@ -70,23 +60,12 @@ function theme() {
         h2heading.style.background = "linear-gradient(90deg, rgba(193, 46, 127, 1) 0%, rgba(224, 29, 253, 1) 47%, rgba(136, 69, 252, 1) 100%)";
         h2heading.style.WebkitBackgroundClip = "text";
         h2heading.style.WebkitTextFillColor = "transparent";
-        for (let i = 0; i < taskboxs.length; i++) {
-            taskboxs[i].style.backgroundColor = "rgb(77, 77, 77)";
-        }
-        for (let i = 0; i < edit_elements.length; i++) {
-            var elements = edit_elements[i];
-            elements.classList.remove('blackedit');
-            elements.classList.add('pinkedit');
-            var box = box_elements[i];
-            box.style.backgroundColor = "rgb(77, 77, 77)";
-            if (elements.innerText.toLowerCase() == "save") {
-                var inp = input_elements[i];
-                inp.style.color = "#e01bd3";
-            }
-        }
+        showtasks();
 
     }
 }
+
+
 
 window.addEventListener('load', () => {
 
@@ -94,10 +73,9 @@ window.addEventListener('load', () => {
     const input = document.querySelector("#task");
     const list = document.querySelector("#tasks");
 
-
+    showtasks();
     forms.addEventListener('submit', (e) => {
         e.preventDefault();
-
 
         const task = input.value;
         if (task.trim() != 0) {
@@ -113,128 +91,15 @@ window.addEventListener('load', () => {
             }
             taskobject.push(task);
             localStorage.setItem("storedtasks", JSON.stringify(taskobject));
-
-            //creating the parent element div task_element 
-            const task_element = document.createElement("div");
-            task_element.classList.add("task");
-            box_elements.push(task_element);
-
-            //creating the first child element div task_content  
-            const task_content = document.createElement("div");
-            task_content.classList.add("content");
-            task_element.appendChild(task_content);
-
-            //creating the child element input of the task_content
-            const input_element = document.createElement("input");
-            input_element.classList.add("text");
-            input_element.type = "text";
-            input_element.value = task;
-            input_element.setAttribute("readonly", "asd");
-            input_elements.push(input_element);
-            task_content.appendChild(input_element);
-
-            //creating the second child element div of the task_element
-            const actions = document.createElement("div");
-            actions.classList.add("actions");
-
-            edit = document.createElement("button");
-            edit.classList.add("edit");
-            edit.innerHTML = "Edit";
-            edit_elements.push(edit);
-
-            if (colorflag == "white") {
-                edit.classList.add('pinkedit');
-                task_element.style.backgroundColor = "rgb(77, 77, 77)";
-            } else {
-                edit.classList.add('blackedit');
-                task_element.style.backgroundColor = "rgb(37 39 60)";
-            }
-
-
-            const deleter = document.createElement("button");
-            deleter.classList.add("delete");
-            deleter.innerHTML = "Delete";
-
-            actions.appendChild(edit);
-            actions.appendChild(deleter);
-
-            task_element.appendChild(actions);
-            list.appendChild(task_element);
             input.value = "";
-
-
-            edit.addEventListener("click", () => {
-                if (edit.innerText.toLowerCase() == "edit") {
-                    if (colorflag == "white") {
-                        input_element.style.color = " #e01bd3";
-                    } else {
-                        input_element.style.color = "#00d9ff";
-                    }
-                    input_element.removeAttribute("readonly");
-                    input_element.focus();
-                    edit.innerText = "Save";
-                }
-                else {
-                    input_element.setAttribute("readonly", "readonly");
-                    input_element.style.color = "white";
-                    edit.innerText = "Edit";
-                }
-
-            })
-
-            // Get the modal element
-            var modal = document.getElementById("modal");
-
-            // Get the <span> element that closes the modal
-            var closeBtn = document.getElementsByClassName("close")[0];
-
-            // Get the modal message element
-            var modalMessage = document.getElementById("modalMessage");
-
-            // Function to display the modal
-            function showModal(message) {
-                modalMessage.textContent = message;
-                modal.style.display = "block";
-            }
-
-            // Function to close the modal with fade-out animation
-            function closeModal() {
-                modal.classList.add("fade-out");
-                setTimeout(function () {
-                    modal.style.display = "none";
-                    modal.classList.remove("fade-out");
-                }, 300); // Wait for the fade-out animation to complete before hiding the modal
-            }
-
-            // Add event listener to the delete button
-            deleter.addEventListener("click", function () {
-                var confirmation = confirm("Are you sure you want to delete?");
-
-                if (confirmation) {
-                    showModal("Item deleted successfully.");
-                    list.removeChild(task_element);
-                    // Code to delete the item
-                    // list.removeChild(task_element);
-                } else {
-                    showModal("Deletion cancelled.");
-                }
-            });
-
-            // Add event listener to the close button
-            closeBtn.addEventListener("click", closeModal);
-
+            showtasks();
         }
         else {
             alert("Please Enter A Task To Add Into List ");
         }
-
-
     });
-
-    
-   
-
 });
+
 
 function showtasks() {
 
@@ -249,55 +114,103 @@ function showtasks() {
     let textfields = document.getElementById("tasks");
     let html = "";
 
-    taskobject.forEach((item,index) => {
+    if(colorflag=="dark"){
+        btnclass="blackedit";
+        inpclass="taskdark"
+    }else{
+        btnclass="pinkedit";
+        inpclass="task"
+    }
+
+    taskobject.forEach((item, index) => {
 
         console.log(index);
-        html += ` <div class="task">
+        html += ` <div class="${inpclass}">
         <div class="content">
-            <input type="text " class="text" value="${item}" readonly>
+            <input type="text" class="text" value="${item}" readonly id="textInput${index}">
         </div>
         <div class="actions">
-            <button class="pinkedit" onclick="editfunc(${index})"> Edit </button>
-            <button class="delete">Delete</button>
+            <button class="${btnclass}" onclick="editfunc(${index})" id="editbtn"> Edit </button>
+            <button class="delete" onclick="deletefunc(${index})">Delete</button>
 
         </div>
         </div>`
     });
+    
     textfields.innerHTML = html;
 };
 
-function editfunc(index){
 
-    let editbox=document.querySelectorAll(".text");
-    let editbtn=document.querySelectorAll(".pinkedit");
+function editfunc(index) {
+
+    let editbox = document.querySelectorAll(".text");
+    let editbtn = document.querySelectorAll('#editbtn');
     let webstorage = localStorage.getItem("storedtasks");
     let taskobject = JSON.parse(webstorage);
     editbox[index].removeAttribute("readonly");
-    editbtn[index].innerHTML="Save";
+    editbtn[index].innerHTML = "Save";
     editbox[index].focus();
-   
-     
-    editbtn[index].addEventListener("click",function(){
-        taskobject[index]=editbox[index].value;
-        localStorage.setItem("storedtasks", JSON.stringify(taskobject));
-        editbtn[index].innerHTML="Edit";
-        editbox[index].setAttribute("readonly","readonly");
+    editbtn[index].addEventListener("click", function () {
+
+        if (editbox[index].value.trim() === '') {
+            alert("Edit Task Field Is Empty ! Add Task To Edit ");
+        } else {
+            taskobject[index] = editbox[index].value;
+            localStorage.setItem("storedtasks", JSON.stringify(taskobject));
+            editbtn[index].innerHTML = "Edit";
+            editbox[index].setAttribute("readonly", "readonly");
+            showtasks();
+        }
     });
 }
 
-function deletefunc(index){
- 
-    let deletedbox=document.querySelectorAll(".text");
-    let deletebtn=document.querySelectorAll(".delete");
+function deletefunc(index) {
+
+    var confirmation = confirm("Are you sure you want to delete?");
+
+  if (confirmation) {
+    showModal("Item deleted successfully.");
     let webstorage = localStorage.getItem("storedtasks");
     let taskobject = JSON.parse(webstorage);
-
-    // deletebtn[index].addEventListener("click",function(){
-    //     taskobject[index]=editbox[index].value;
-    //     localStorage.setItem("storedtasks", JSON.stringify(taskobject));
-    //     deletedbox[index].remove;
-
-    // })
+    taskobject.splice(index, 1);
+    localStorage.setItem("storedtasks", JSON.stringify(taskobject));
+    showtasks();
+  } else {
+    showModal("Deletion cancelled.");
+  }
+    
 
 }
+
+
+// Get the modal element
+var modal = document.getElementById("modal");
+
+// Get the <span> element that closes the modal
+var closeBtn = document.getElementsByClassName("close")[0];
+
+// Get the modal message element
+var modalMessage = document.getElementById("modalMessage");
+
+// Function to display the modal
+function showModal(message) {
+  modalMessage.textContent = message;
+  modal.style.display = "block";
+}
+
+// Function to close the modal with fade-out animation
+function closeModal() {
+  modal.classList.add("fade-out");
+  setTimeout(function() {
+    modal.style.display = "none";
+    modal.classList.remove("fade-out");
+  }, 300); // Wait for the fade-out animation to complete before hiding the modal
+}
+
+// Add event listener to the delete button
+
+
+// Add event listener to the close button
+closeBtn.addEventListener("click", closeModal);
+
